+++ src/common.ts
@@ -1,9 +1,11 @@
 export type Metadata = Record<string, any>;
 
 export type Config = {
+    outputFolder?: string,
     eleventy?: boolean,
     defaultStyles?: boolean,
     customStyles?: string,
+    templatesFolder?: string,
 }
 
 export class Result extends String {
@@ -13,3 +15,9 @@ export class Result extends String {
         this.metadata = {};
     }
 }
+
+export function RegExpBuilder(regex: string, flag: string = 'mgi') {
+    return RegExp(regex.replace(/ /g, '').replace(/\|\|.+?\|\|/g, ''), flag);
+}
+
+export const RawString = String.raw;