+++ src/compile.d.ts
@@ -1,2 +1,0 @@
-import { Config } from './types';
-export declare function compile(dir?: string, config?: Config): void;