+++ src/compile.ts
@@ -1,12 +1,9 @@
 const fs = require('fs');
 const glob = require('glob');
 const formatter = require('html-formatter');
-const novasheets = require('novasheets');
 
 import { parse } from './parse';
-import { Config, Result } from './types';
-
-const OUT_FOLDER = 'wikity-out/';
+import { Config, Result, RegExpBuilder as re, RawString as r } from './common';
 
 export function compile(dir: string = '.', config: Config = {}): void {
     let stylesCreated = false;
@@ -14,11 +11,12 @@ export function compile(dir: string = '.', config: Config = {}): void {
     const files = glob.sync((dir || '.') + "/**/*.wiki", {});
     files.forEach((file: string) => {
         let data: string = fs.readFileSync(file, { encoding: 'utf8' });
-        let content: Result = parse(data);
+        let content: Result = parse(data, config);
         let outText: string = content.toString();
+        const templatesFolder = config.templatesFolder || 'templates';
 
-        let [, folder, filename]: string[] = file.match(/^(.+?[\/\\])((?:templates[\/\\])?[^\/\\]+)$/) as RegExpMatchArray;
-        let outFolder: string = (dir || folder || '.') + '/' + OUT_FOLDER;
+        let [, folder, filename]: string[] = file.match(re(r`^(.+?[\/\\]) ((?:${templatesFolder}[\/\\])?[^\/\\]+)$`,'')) as RegExpMatchArray;
+        let outFolder: string = (dir || folder || '.') + '/' + (config.outputFolder || 'wikity-out') + '/';
         let outFilename: string = filename.replace(/ /g, '_').replace('.wiki', '.html');
         let url: string = outFilename.replace(/(?<=^|\/)\w/g, m => m.toUpperCase())
         let displayTitle: string = content.metadata.displayTitle || url.replace('.html', '');
@@ -95,7 +93,7 @@ export function compile(dir: string = '.', config: Config = {}): void {
         stylesCreated = true;
         let styles: string = '';
         if (config.defaultStyles !== false) {
-            styles += novasheets.parse(`
+            styles += `
                 body {font-family: sans-serif; margin: 4em; max-width: 1000px; background: #eee;}
                 main {margin: 3em -1em; background: #fff; padding: 1em;}
                 h1, h2 {margin-bottom: 0.6em; font-weight: normal; border-bottom: 1px solid #a2a9b1;}
@@ -104,15 +102,19 @@ export function compile(dir: string = '.', config: Config = {}): void {
                 dd, dl dl {margin-block: 0; margin-inline-start: 30px;}
 
                 a:not(:hover) {text-decoration: none;}
-                a.internal-link {color: #04a;} &:visited {color: #26d;}
-                a.external-link {color: #36b;} &:visited {color: #58d;} &::after {content: '\\1f855';}
-                a.redlink {color: #d33;} &:visited {color: #b44;}
+                a.internal-link {color: #04a;}
+                a.internal-link:visited {color: #26d;}
+                a.external-link {color: #36b;}
+                &:visited {color: #58d;}
+                a.external-link&::after {content: '\\1f855';}
+                a.redlink {color: #d33;}
+                a.redlink:visited {color: #b44;}
 
                 #toc {display: inline-block; border: 1px solid #aab; padding: 8px; background-color: #f8f8f8; font-size: 95%;}
-                &-heading {display: block; text-align: center;}
-                & ol {margin: 0 0 0 1.3em;}
-                &.toc-hidden {height: 1em;} % ol {display: none;}
-            `.replace(/^\s+/gm, ''));
+                #toc-heading {display: block; text-align: center;}
+                #toc ol {margin: 0 0 0 1.3em;}
+                #toc.toc-hidden {height: 1em;} % ol {display: none;}
+            `.replace(/^\s+/gm, '');
         }
         if (config.customStyles) styles += config.customStyles;
         fs.writeFileSync(outFolder + 'wiki.css.njk', '---\npermalink: /wiki.css\n---\n' + styles);