+++ src/index.d.ts
@@ -1,6 +1,0 @@
-import { compile } from './compile';
-declare const _default: {
-    compile: typeof compile;
-    parse: (data: string) => string;
-};
-export = _default;