+++ src/parse.d.ts
@@ -1,2 +1,0 @@
-import { Result } from './types';
-export declare function parse(data: string): Result;