+++ test/.eleventy.js
@@ -1,2 +1,5 @@
-const wikity = require('../src/index.js');
-wikity.compile('.', { eleventy: true });
+const wikity = require('../src/index.js')
+
+module.exports = function (eleventyConfig) {
+    eleventyConfig.addPlugin(() => wikity.eleventyPlugin());
+}